crowditServer/requirements.txt
Django==1.4
MySQL-python==1.2.3
WebOb==1.1.1
boto==2.5.2
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
dj-database-url==0.2.1
django-oauth-plus==2.0
django-social-auth==0.7.1
django-storages==1.1.5
django-tastypie==0.9.11
httplib2==0.7.4
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
requests==0.10.2
simplejson==2.6.0
virtualenv==1.7.2
virtualenv-clone==0.2.4
virtualenvwrapper==3.5
wsgiref==0.1.2
